Graphic improvement Added icon, window title and preset text to input edit label.
